feat: extract descriptor set and pipeline modules
