+++ src/gpu/vulkan/examples.rs
@@ -1,0 +1,1 @@
+//! Example utilities have been moved to integration tests and the `examples/` directory.
